+++ CRUDA/db/Scripts/cruda.TransactionsCreateTable.sql
@@ -1,7 +1,7 @@
 ﻿IF (SELECT object_id('[cruda].[Transactions]', 'U')) IS NOT NULL
     DROP TABLE [cruda].[Transactions]
 CREATE TABLE [cruda].[Transactions]([Id] [int] IDENTITY(1,1) NOT NULL
-                                   ,[LoginId] [bigint] NOT NULL
+                                   ,[LoginId] [int] NOT NULL
                                    ,[IsConfirmed] [bit] NULL
                                    ,[CreatedAt] datetime NOT NULL
                                    ,[CreatedBy] varchar(25) NOT NULL
